move pending part to a module in the crate
