+++ src/middlewares/mod.rs
@@ -6,3 +6,4 @@ pub use commit_id_schema::{
     decode_history_number_rev, encode_history_number_rev, CommitID, CommitIDSchema, HistoryNumber,
 };
 pub use key_value_store_bulks::{ChangeKey, KeyValueStoreBulks};
+pub use versioned_flat_key_value::PendingError;
