--- src/pending_part/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "in-memory-tree"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-slab = "0.4.9"
-thiserror = "1"
-
-[dev-dependencies]
-rand = "0.8.0"
-rand_distr = "0.4.0"
